refactor: change domain classes' access modifiers to default
