+++ src/main/java/lotto/domain/LottoTickets.java
@@ -7,7 +7,7 @@ import java.util.Map;
 public class LottoTickets {
     private final List<LottoTicket> lottoTickets;
 
-    public LottoTickets(final List<LottoTicket> lottoTickets) {
+    LottoTickets(final List<LottoTicket> lottoTickets) {
         this.lottoTickets = lottoTickets;
     }
