+++ src/test/java/lotto/domain/LottoTicketsFactoryTest.java
@@ -1,8 +1,7 @@
-package lotto.service;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import lotto.domain.LottoTickets;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
